Refactor HTTP components into modular architecture
• Move server logic to src/http/server.rs
• Move client logic to src/http/client.rs
• Extract CLI handling to src/http/cli.rs
• Add shared utilities in src/http/common.rs
• Simplify binaries to minimal instantiation code
