+++ Cargo.toml
@@ -4,6 +4,10 @@ version = "0.1.0"
 edition = "2024"
 license = "MIT"
 
+[features]
+default = ["http"]
+http = []
+
 [dependencies]
 axum = "0.8"
 tokio = { version = "1.47.1", features = ["full"] }
