+++ tests/producer_tests.rs
@@ -1,7 +1,7 @@
 mod test_helpers;
 
 use flashq::Record;
-use flashq::api::*;
+use flashq::http::*;
 use test_helpers::{TestClient, TestServer};
 
 #[tokio::test]
